Removed unused code; optimized imports; added states to watershed list item

+++ app/src/main/java/com/viableindustries/waterreporter/data/TerritoryProperties.java
@@ -28,7 +28,7 @@ public class TerritoryProperties implements Serializable {
     public String huc_6_name;
 
     @SerializedName("huc_8_code")
-    public long huc_8_code;
+    public int huc_8_code;
 
     @SerializedName("huc_8_name")
     public String huc_8_name;
